+++ modules/apim-adapter/src/main/java/com/axway/apim/api/API.java
@@ -460,7 +460,7 @@ public class API implements CustomPropertiesEntity {
 
 	@Override
 	public String toString() {
-		return this.getClass().getSimpleName() + " [path=" + path + ", id (FE-API)=" + id + ", apiId (BE-API)=" + apiId + "]";
+		return this.getClass().getSimpleName() + " [path=" + path + ", id (FE-API)=" + id + ", apiId (BE-API)=" + apiId + ", vhost=" + vhost + ", apiRoutingKey=" + apiRoutingKey + "]";
 	}
 
 	public String getApiDefinitionImport() {